requirements.txt
cudatoolkit == 9.0
cudnn == 7.3.1
tensorflow-gpu == 1.12.0
opencv-python == 4.0.0.21
scikit-image == 0.14.2
scikit-learn == 0.20.2
scipy == 1.1.0
tqdm == 4.28.1